+++ love/src/redux/selectors/selectors.js
@@ -766,15 +766,15 @@ export const getATMCSState = (state) => {
 export const getAuxiliaryTelescopeState = (state) => {
   const subscriptions = [
     'telemetry-ATPtg-0-mountStatus',
-    'telemetry-ATPtg-0-mountPosition',
+    'telemetry-ATPtg-0-mountPositions',
     'event-ATPtg-0-currentTarget',
   ];
   const data = getStreamsData(state, subscriptions);
   return {
     telescopeRAHour: data['telemetry-ATPtg-0-mountStatus']?.mountRA?.value ?? 0,
-    telescopeRADeg: data['telemetry-ATPtg-0-mountPosition']?.ra?.value ?? 0,
+    telescopeRADeg: data['telemetry-ATPtg-0-mountPositions']?.ra?.value[0] ?? 0,
     telescopeDecDeg: data['telemetry-ATPtg-0-mountStatus']?.mountDec?.value ?? 0,
-    telescopeRotatorRad: data['telemetry-ATPtg-0-mountStatus']?.mountRot?.value ?? 0,
+    telescopeRotatorDeg: data['telemetry-ATPtg-0-mountStatus']?.mountRot?.value ?? 0,
     targetName: data['event-ATPtg-0-currentTarget']?.[0].targetName?.value ?? 'Unknown',
   };
 };
@@ -1318,7 +1318,7 @@ export const getMainTelescopeState = (state) => {
     telescopeRAHour: data['telemetry-MTPtg-0-mountStatus']?.mountRA?.value ?? 0,
     telescopeRADeg: data['telemetry-MTPtg-0-mountPosition']?.ra?.value ?? 0,
     telescopeDecDeg: data['telemetry-MTPtg-0-mountStatus']?.mountDec?.value ?? 0,
-    telescopeRotatorRad: data['telemetry-MTPtg-0-mountStatus']?.mountRot?.value ?? 0,
+    telescopeRotatorDeg: data['telemetry-MTPtg-0-mountStatus']?.mountRot?.value ?? 0,
     targetName: data['event-MTPtg-0-currentTarget']?.[0].targetName?.value ?? 'Unknown',
   };
 };