+++ CHANGELOG.rst
@@ -2,6 +2,17 @@
 Version History
 ===============
 
+v6.7.0
+------
+
+* Fix telescope RA and Rotator display on DomeSummaryTable components. `<https://github.com/lsst-ts/LOVE-frontend/pull/685>`_
+* Add a button to log alarms from the AlarmsList and AlarmsTable components. `<https://github.com/lsst-ts/LOVE-frontend/pull/684>`_
+
+v6.6.0
+------
+
+* Extend LOVE Jira interface to comply with new OBS systems hierarchy. `<https://github.com/lsst-ts/LOVE-frontend/pull/683>`_
+
 v6.5.0
 ------